refactor: pull book search code up to route file for readability
